feat: test-mail action, unified email layout, expanded variables, decimal percentages
- Add per-row "Test mail naar mij sturen" action on PopupFollowUpFlow Repeater that builds a transient PopupFollowUpEmail + PopupView from the current form-state and sends synchronously via Mail::to()->sendNow().
- Refactor PopupFollowUpMail to render through the unified dashed-core::emails.layout (same wrapper, header, footer, primary-color band as all other system mails). Block types are mapped per-block to <tr><td> rows that fit the layout's table structure; popup-specific types (paragraph/usp/discount) get aliases or inline rendering.
- Expand variable system to 5 first-class variables (:siteName:, :email:, :discountCode:, :discountValue:, :siteUrl:) substituted in subject and every text/url/label/code field of every block. helperText shows the variables on each individual block field plus on the Builder. createDefault() seeds use of all variables across the 3 standard steps and adds a button-block per step linking to :siteUrl:.
- Discount-block now displays the actual discount value ("Bespaar 10%" or "Bespaar EUR 5,00") derived from DiscountCode->type + discount_percentage / discount_amount via new resolvePopupDiscountInfo(). Percentage formatting strips trailing zeros (10%, 12,5%, 7,75%).
- Decimal discount percentages: dashed__popups.discount_percentage and dashed__popup_variants.discount_percentage_override migrated from integer to decimal(5,2). Filament inputs have ->step(0.01). Casts updated to decimal:2. resolvedDiscountPercentage() now returns float.
- PopupFollowUpMail constructor: fix Unknown Property collision with parent Mailable::locale by removing readonly from the promoted property and assigning to the parent property in the body.
- PopupFollowUpFlowResource Repeater (emails) gets a mutateRelationshipDataBeforeFillUsing callback that unwraps Spatie's translatable JSON before filling the Builder, fixing the "Undefined array key 'type'" crash.
- New public properties PopupFollowUpMail::$previewDiscountCode and $previewDiscountValue used by the test-mail action so admins see variable substitution working without an actual popup conversion.
- Strip em-dashes from all source files, blade templates and CHANGELOG entries in this package.

+++ src/Livewire/Popup.php
@@ -214,7 +214,7 @@ class Popup extends Component
         $codePrefix = $variant?->code_prefix
             ? 'WELKOM-'.strtoupper($variant->code_prefix).'-'
             : 'WELKOM-';
-        $discountPercentage = $variant?->resolvedDiscountPercentage() ?? (int) $this->popup->discount_percentage;
+        $discountPercentage = $variant?->resolvedDiscountPercentage() ?? (float) $this->popup->discount_percentage;
         $validDays = $variant?->resolvedValidDays() ?? (int) $this->popup->discount_valid_days;
 
         $existingCartWithCode = Cart::where('abandoned_email', $email)
